+++ plugins/grafana/src/components/page/Page.tsx
@@ -1,0 +1,70 @@
+import {
+  Drawer,
+  DrawerContent,
+  DrawerContentBody,
+  PageSection,
+  PageSectionVariants,
+  Title,
+} from '@patternfly/react-core';
+import React, { useEffect, useState } from 'react';
+import { useHistory, useLocation } from 'react-router-dom';
+
+import Dashboards from './Dashboards';
+import { IOptions } from '../../utils/interfaces';
+import { IPluginPageProps } from '@kobsio/plugin-core';
+import PageToolbar from './PageToolbar';
+import { getOptionsFromSearch } from '../../utils/helpers';
+
+const Page: React.FunctionComponent<IPluginPageProps> = ({
+  name,
+  displayName,
+  description,
+  options,
+}: IPluginPageProps) => {
+  const location = useLocation();
+  const history = useHistory();
+  const [pageOptions, setPageOptions] = useState<IOptions>(getOptionsFromSearch(location.search));
+
+  // changePageOptions is used to change the options to get a list of dashboards from Grafna. Instead of directly
+  // modifying the options state we change the URL parameters.
+  const changePageOptions = (opts: IOptions): void => {
+    history.push({
+      pathname: location.pathname,
+      search: `?query=${encodeURIComponent(opts.query)}`,
+    });
+  };
+
+  // useEffect is used to set the options every time the search location for the current URL changes. The URL is changed
+  // via the changePageOptions function. When the search location is changed we modify the options state.
+  useEffect(() => {
+    setPageOptions(getOptionsFromSearch(location.search));
+  }, [location.search]);
+
+  return (
+    <React.Fragment>
+      <PageSection variant={PageSectionVariants.light}>
+        <Title headingLevel="h6" size="xl">
+          {displayName}
+        </Title>
+        <p>{description}</p>
+        <PageToolbar name={name} query={pageOptions.query} setOptions={changePageOptions} />
+      </PageSection>
+
+      <Drawer isExpanded={false}>
+        <DrawerContent panelContent={undefined}>
+          <DrawerContentBody>
+            <PageSection style={{ minHeight: '100%' }} variant={PageSectionVariants.default}>
+              <Dashboards
+                name={name}
+                query={pageOptions.query}
+                publicAddress={options && options.publicAddress ? options.publicAddress : ''}
+              />
+            </PageSection>
+          </DrawerContentBody>
+        </DrawerContent>
+      </Drawer>
+    </React.Fragment>
+  );
+};
+
+export default Page;